Add Startup script to workaround backup camera issues on pi
This script will start and stop the program multiple times, to clear away the error.

The Backup Camera gives us an error when starting for the first time:
`libv4l2 error setting pixformat input/output error`
And then floods output with `HIGHGUI` errors.

The issue goes away after the Backup Camera is stopped and restarted.

I tried to update opencv libraries, upgrade the pi's libraries, increase the swap size, and change the resolution on the pi to no avail. This is a consistent workaround, that opens properly 10/10 times.